added new scheduler module
The purpose of schedulers ist submit or start jobs either via shell or an interface to a queeing system like PBS
or SLURM. For now only the abstract scheduler base class defining the interface and a very basic scheduler to start jobs locally via shell are implemented.
Also a factory class for creating these schedulers has been implemented.

PBS and SLURM based schedulers will be added in the future