Added: New iterator for Bayesian optimization.
The iterator wraps the gpflowopt library and is currently able to
harness the plain vanilla BO capabilities of gpflowotp.
The covariance function of the GP used to perform BO is hard-coded to
Mater52.

Added : Rosenbrock test function to the set of exemplary simulator
funtions.

Adopted: Iterator base class to accmodate for new BO iterator.