Added new class hierarchy called regression_approximations.
These so-called regression approximations are essentially regression
models, whihc are not called models within QUEENS to avoid confusion
with simulation models and the like.

Currently, there is only one actual implementation of a regression
approximation, a Gaussian process regression approximation based on the
GPy library. The type of covariance function used for the GP is
currently still hard coded, but will be made an option in the future.

The newly added regression_approximations will be the fundamental
building block for soon to be introduced surrogate models.