Refactored and cleaned up scheduler classes and re-enabled the ability to perform computations on Kaiser.
This involved implementation of new class methods for the schedulers and
some shifting as to hwo some options are passed down the the actual
simulation driver. The old scheduler factory is deleted. Both the
resource class and the job_interface class needed some minro
modification in order to work with the new structure.

The job_interface test case needed to modified as well to acommodate the
new interface of the schedulers.

The new functionality can be tested with the input file
baci_kaiser_condig. The old try_out_job_scheduler script is removed.