Call local scheduler & driver with same Python
Typically, we want the same Python executable for QUEEN's main and the
drivers. It is recommended to ensure this by running QUEENS within an
active, QUEENS-specific conda environment.
However, calling QUEENS with a specific Python executable should also
work. For example, the Pycharm IDE does not activate the environment,
but runs Python files by specifying the absolute path to the python
executable of the specified environment.
In this case, calling "python" in a subprocess.Popen (like in the
local_scheduler) will result in the attempt to use system's default
Python executable (since the PATH variables are not updated). Naturally
this almost surely results in errors.
By calling with "sys.executable" we ensure that the same interpreters
are being used whether the environment is activated or not.
Therefore one can run the local driver from within Pycharm.