Enable vector-valued model responses (#20)
The requirement for low level model responses, i.e. for example
simulator functions, all drivers and approximations, is that it is
array_like.
The array_like response will then be converted at interface level (or at
approximation level in the case of the approximation interface) to an
ndarray. The rows (first dimension) of this ndarray corresponds to
individual model evaluations, the remaining dimensions (at least 1)
correspond to the model dimensions.
The difference between scalar-valued and vector-valued models is the
size of the second dimension.
In case of matrix-valued responses there would be a third dimension.