Add method to retrieve pre-calculated model response (#25)
The solution algorithms in the scipy.optimize (and most other
optimization toolboxes) do not exploit the potential for parallel
objective function evaluation.

This feature is intended to circumvent this disadvantage:
during an iteration of the solution algorithm we plan to pre-calculate
all function evaluations in parallel as soon as the evaluation points are
theoretically known (this potentially implies some implementation effort
per solution algorithm).
The solutions are cached/ stored and retrieved from the storage when the
solution algorithm asks for it later in the iteration.
This should be a lot faster than serially evaluating the function
serially.