Finite difference approximation of Jacobian (#25)
fd_jacobian basically is a refactoring/ reimplementation of parts of
the scipy.optimize._numdiff module.

The idea is to separate the computation of all positions necessary for
the finite difference (FD) approximation from the actual evaluation of
the function at these positions.
The goal of this separation is to enable the parallel (simultaneous)
evaluation of the function at all positions which results in significant
speed ups for expensive functions.

The approach is as follows:
1.) Compute positions based on FD scheme and bounds
2.) Evaluate function at these positions (see 1.) (in parallel)
3.) From computed function values and positions approximate Jacobian