Multiple Chains in MH Iterator (#27)
Enable multiple, parallel Markov chains.
This is realised by adding an axis to the respective numpy.ndarrays.

The axes now correspond to:
  axis 0: sample ID
  axis 1: chain ID
  axis 2: variables/sample position

While this is certainly not the most efficient implementation of the
embarrassingly parallelism, this implementation already gives a
significant speed-up (compared to running the chains sequentially).

Multiple chains can be used to calculate diagnostics of the chains.
This is realised by ArviZ package.

When the MH iterator is used as the MCMC kernel of a SMC algorithm,
parallel chains can be used as well.

utils and tests have been updated accordingly.