Fix matplotlib backend issue (#27)
If matplotlib has already been imported in headless mode (backend: Agg),
trying to change to TkAgg backend results in an error.

This occurs now because the Metropolis-Hastings iterator also uses
matplotlib (through ArviZ) (introduced with 4d5c620f).
In iterator.py it MH interator is imported before the MorrisSALib
iterator (alphabetical order). The MorrisSALib iterator tries to set
the matplotlib backend to 'TkAgg', after it has already been set by the
imports of ArviZ and matplotlib in MH iterator.
The problem only occurs on systems without X-server (i.e., $Display).
This is for example the case when you run QUEENS via ssh without -X or
in our Gitlab testing setup (the GitLab Runner).
In MH iterator, ArviZ now seems to automatically set the headless Agg
backend, which can not be changed by MorrisSALib.

Solution: check for the headless 'Agg' backend before trying to set the
backend to 'TkAgg' in the MorrisSALib Iterator.