Make draw_trace optional, add test, refactor (#26)
New Feature:
make the plotting of the trace optional. Setting the new input
parameter plot_trace_every=0 will disable it completely during runtime.
Otherwise the current trace will be plotted every plot_trace_every-th
step.
The final plotting (i.e., of the results) can be turned of in the
results description.

Test:
Added a new test case for generic tempering which is based on a
multivariate Gaussian with weak covariance structure (not diagonal but
small condition).

Refactor:
- Blackened the Sequential Monte Carlo Iterator
- Turned off the plotting of traces in the test case except for the
  results of the test_gaussian_smc (to include it in a test).