Improve tempering functions, add test (#26)
The special cases of the tempering are now handled rigorously. Namely
those are when the temperature (the tempering parameter) is either zero
or one.
The reasoning behind this is that in these cases the respective
distributions are clearly indicated by the temperature.
A temperature of zero refers to the initial distribution.
A temperature of one indicates that the goal distribution is reached.

Additionally, I chose to exclude the case of positive infinite logpdfs.
The reasoning behind is that the results of (+inf + -inf) is ambiguous.
To circumvent the a tedious management of all special cases, the
unlikely case of logpdf=+inf is excluded.
Note that logpdf=-inf on the other hand is very likely. Consider for
example the uniform distribution is zero (log(0)=-inf) in many areas.

Added unittests for all tempering functionalities.