Introduce pylint with a basic .pylintrc-file (#9)
From Pylint docs:
"Pylint is a tool that checks for errors in Python code, tries to
enforce a coding standard and looks for code smells.  It can also look
for certain type errors, it can recommend suggestions about how
particular blocks can be refactored and can offer you details about the
code's complexity."

It is fully cusomisable via the .pylintrc-file.
The present .pylintrc is the default except for the max-line-length
which has been increased to 99.
It can be created with:
  pylint --max-line-length=99 --generate-rcfile > .pylintrc