Adjust tests for BaciDriverSchmarrn (#39)
Get the test of the old dummy_driver_baci_pbs_kaiser running for the new
BaciDriverSchmarrn.

mpi_run command is not needed anymore in the driver (it is hidden in the
singularity image now) and therefore does not need to be tested anymore.

Moved test suite of BaciDriverSchmarrn to pytest framework.

Introduced a base_settings fixture that might need some rework but works
for now to be used as input for Driver class fixtures.

Slightly refactored the conftest of the respective tests.


Former-commit-id: a259baca39c52f86cb5f1bc2a3ea47a7074a182e