Added dummy function to make test run again (#39)
This is an intermediate step to make the test suite at least run again.
The test of course fails for now and the dummy function need to be
deleted and replaced with the respective new implemented features of the
Driver class.
All tests then have to be adjusted to test the respective new
implementation.


Former-commit-id: cf0a3e0b3a587e6c869b7ed6fddfe584edd4d6f4