Add testpaths variable to setup.cfg (#39)
We have been picking up undesired tests (e.g. in the setup_remote.py).
This leads to skipped pseudo-tests and potentially even errors that fail
the complete suite (see the errors in Pipeline #148268).
This is exactly what the testpaths variable is for (from pytest's doc):

Sets list of directories that should be searched for tests when no
specific directories, files or test ids are given in the command line
when executing pytest from the rootdir directory. Useful when all
project tests are in a known location to speed up test collection and to
avoid picking up undesired tests by accident.


Former-commit-id: afb7f661f0117acd540cb87fdecd1127c6def51e