Re-enable bruteforce as computation resource
Based on the implementation for Deep, this commit fixes the relevant
bruteforce files.
Notably, the computation takes place on the compute nodes (realised by
the new jobscript template). To reduce network traffic, the results are
written in a working directory (workdir, "scratch") on the compute node.
Therefor, QUEENS' bruteforce driver need to know the workdir variable
(see changes in baci_driver_bruteforce.py).
It follows directly that we have to supply the workdir as an
input-parameter to the singularity image.
IMPORTANT NOTE: at this point, we exploit the structure of BACI calls on
the cluster(see the donottouch.sh, which should work for both (BACI and
QUEENS, together with the newly added jobscript_slurm_scheduler.sh).
post_drt_monitor as post-processing filter may currently only run on the
login/ master node, therefore the results have to be copied to the
user's home first (compare second call of executable in
jobscript_slurm_scheduler.sh with the --post=true flag
(first call is hidden in RunProgram)).

Fix a minor bug within the slurm_scheduler, where a necessary dictionary
was not declared.

Update the example input file to incorporate/ account for the above
changes. To run the input file update user specific paths
e.g. the path to your BACI executable on bruteforce.


Former-commit-id: ec060a19f163f9421ee6fd82a7045de519c5be33