Fix BACI post-processing call in jobscript slurm
Due to an error in donottouch_new.sh on bruteforce, the post-processing
call at the end of the jobscript was done from a non existing directory.
Therefor, a call with $MPI_RUN was not possible, causing problems in the
actual call of e.g. post_drt_monitor within the singularity image.

This is now fixed in donottouch_new.sh and jobscript.
The post-processing call should work as expected.
The only drawback is that we may only post process with a single core
at the moment (because post_drt_monitor can only be called with -np 1).
This restriction could be easily dropped if needed.


Former-commit-id: 5def1517d29d771a6baa6aff851d6133f0733c6a