Fix Python 3.8 Syntax Warning "is" with literal.
Of course fix also:
"is not" with literal.

Why?
From Python 3.8 release note
https://docs.python.org/3.8/whatsnew/3.8.html#changes-in-python-behavior

The compiler now produces a SyntaxWarning when identity checks
(is and is not) are used with certain types of literals
(e.g. strings, numbers).
These can often work by accident in CPython, but are not guaranteed
by the language spec.
The warning advises users to use equality tests (== and !=) instead.


Former-commit-id: b138473b310ce7044a23825243813e162862a87e