Fix usage of args in remote_main
the args supplied to to main in remote_main was not the actual list of
arguments but only the first one (see l. 66).
Since we called parse_args() without an argument it under hood got the
right list of arguments anyways but did not use the supplied args.
The fix here does the same as was previously happening under the hood
but makes it visible and in line so that the args list before parsing
actually contains the complete list of arguments.


Former-commit-id: 5c46caf65fdf32389ac397dacf3bf0e727f4c3dc