Require astroid>2.3.3 in environment.yml
astroid is a dependency of pylint.
The update to pylint 2.5.0 does not to work with the default
astoird=2.3.3. that is currently installed via the default conda
channel. This leads to erros calling pylint both locally and in the
pipelines.
The error gets fixed by using pyling 2.5.0 with a newer astroid version
e.g. 2.4.0.

Therefore, add the explicit requirement of astroid>2.3.3 for
now.

This could later be reverted if the default conda channel supplies a
higher astroid version.


Former-commit-id: fabc719c1d67ed179e8660560cdd1adc3c2a081f