Make address class variable remove from interface
The address of the database was calculated in both in the JobInterface
and in the actual database class.
It was only used for printing information in the JobInterface.

Since this is just unnecessary code duplication that could lead to
errors in the future -> make the db_address a class variable.