Omit deleting of temporary local jobfile
In the inject routine, we open the temporary local jobfile with the flag
"w" which causes it to be truncated to zero.
Therefore we do not need to delete the file after usage (only to
recreate it for the next job).

This is a core piece of code because it is called for every job.
Omitting this unnecessary operation results in a permanent speed-up for
every QUEENS user in every QUEENS run for every job.