build: readd .git-blame-ignore-revs
When added to the local git config, the file can be used to ignore the
revs it list in a call to git blame.
It contains revs to commits that most likely do not contain the
information one looks for with git blame.
Usually one looks for change of code functionality instead of for
example code style such that for example an automated change in code
style would be a commit to add to the .git-blame-ignore-revs file.

See also the explanations in the file itself.
It can be added to the config with the following command:
git config blame.ignoreRevsFile .git-blame-ignore-revs