refactor(gitignore): Generalize .gitignore of pytest-cov, vscode and .egg-info files/dirs
pytest-cov generates a lot of files next to the `.coverage` file which all start with `.coverage*`.
These files are temporary machine and version specific and should never
be commited. Hence they are added to .gitignore.

The vscode files where ignored based on a specific project name. This is
generalized to ignore the files independent of the projects name.

Ignore all `*.egg-info` directories not only `pqueens.egg-info`.