ci(codechecks): restrict pylint in pipelines to a few warnings for now and run pylint last
The current codebase is nowhere near pylint conformity.
Therefore conformity will be approached over time by enforcing
new and changed files to conform in future commits.
This is achieved by enforcing all warnings and errors in the pylint
precommit hook but only a very small subset in the ci pipelines.

In the future, the whole codebase should be tested against all warnings.