feat(tests): automate markers for tests
There were a couple of tests that had no marker.
The marker had to be added manually at least per file.
This is not necessary since the markers represent the folder structure.

Now all tests in one folder get the correct marker automatically
(because they are in the respective folder).

This mean e.g. that all tests in the "pqueens/tests/unit_tests" folder
now are marked with the marker "unit_tests".