feat(): remove cluster_bind from input options and cleanup benchmarks
the bind variables are defined in the new ClusterConfigs dataclasses and
can easily be moved to a config file in the future.

Removing the cluster_bind from input options also meant that they could
be removed from the cluster test fixtures which invoked a little bit of
a rework.

At this point, I also cleaned up the conftest of benchmark tests (which
turned out to not be needed at all) and some of the issue of the
benchmark tests (these changes are part of #381)