feat: automate handling of QUEENS directories
All data of a QUEENS experiment is now collected in an automatically
created subfolder of a base directory - both for a local run and a
remote run.
Currently, the base dir is:
$HOME/queens-experiments/

The basic structure is as follows:

base-directory/
├─ experiment_1/
│  ├─ job_1
│  ├─ job_2
│  ├─ job_3
├─ experiment_2/
│  ├─ job_1
│  ├─ job_2
├─ singularity_image.sif

This base directory is defined in a central place: config_directories.py
Currently, this base directory is enforced to be in the home of the user.
In the future, it can be made more flexible by moving the definition
of the base directory into a config file.

For the test suite it is already overwritten to a pytest-temp folder for
each test (technically each test has its own base dir).
To prevent the singularity image to be built in each individual test -
it would be expected in the base dir - the singularity path is mocked to
the standard base dir path of the user.

BREAKING CHANGE: experiment_dir can no longer be supplied via QUEENS input-files.