feat: automate handling of QUEENS directories #2
All data of a QUEENS experiment is now collected in an automatically
created subfolder of a base directory - both for a local run and a
remote run.
Currently, the base dir is:
$HOME/queens/

The basic structure is as follows:

base-directory
├─ experiments
|  ├─ experiment_1
|  │  ├─ job_1
|  │  ├─ job_2
|  │  ├─ job_3
|  ├─ experiment_2
|     ├─ job_1
|     ├─ job_2
├─ tests
|  ├─ test_1
|  ├─ test_2
├─ singularity_image.sif
├─ temporary_submission_script.sh
├─ temp.sh

The base directory and (almost?) all other directories are defined in a
central place:
pqueens/utils/config_directories.py

Currently, the base directory is enforced to be in the home of the user.
In the future, it can be made more flexible by moving the definition
of the base directory into a config file.

For the test suite it is already overwritten to a pytest-temp folder for
each test such that technically each test has its own base dir.
This is inline with how pytest suggests to structure test data.
To prevent the singularity image to be built in each individual test -
it is expected in the base dir which as stated above changes for each
test - the singularity path is mocked to the standard base dir
of the user.

The current (and only) singularity_image.sif can locally
always be found under:
$HOME/queens/singularity_image.sif

With this commit, no data should be written to the queens source folder
during a run.

The suggested structure above is not final and up for discussion.
It is implemented in such a way that it can be adapted more easily in
the future.

BREAKING CHANGE: experiment_dir can no longer be set in the input-file