fix: reintroduce working_dir, fix dest_dir
the working directory is the directory where the output is written to
during the actual computation (call of the forward solver).
It is different from the output_dir since the output might be first
written on the computing node itself in order to reduce network and/or
infiniband traffic. In this case working_dir is on the computing node
and the data is expected to be copied to output_dir after computation.
If no working_dir is supplied, it is set equal to the output_dir.

The Dest_DIR is equal to the output_dir, a subdir of the job_dir which
is a folder in the experiment_dir for one job only