feat: add pylint config file for pipeline
The .pylintrc_ci file is identical to the file for local development
.pylintrc with the exception that all messages that pylint currently
throws are disabled.

This enables us to ensure that no additional messages will be added in
the future while at the same time work on the disabled messages one at a
time.

Once all occurrences of a particular message type are fixed, this
message can and should be removed from the disable list in .pylintrc_ci.