ci: simplify-rules-for-jobs-execution
UBUNUTU_TESTS are enabled by default and have to actively be disabled.
If not wanted set "UBUNTU_TESTS 0".

On the main branch, now only the respective tests should be run
(removed running both develop and main tests on default branch)

While the conveniences is reduced a little by having to turn of the
UBUNTU_TESTS explicitly, this simplifies the gitlab-ci.yml a lot.