ci(): fix path to singularity image
There was a mismatch of the path to the singularity image from the code
and the pipeline.
Most likely since the path was not updated accordingly.

For this reason the path is now read from the same place everywhere,
that is, from the config_directories.py file where it is defined once
and used everywhere.

It is needed for
- normal queens runs
- local pytest runs
- pipeline pytest runs (which prebuild the image)