feat: adapt folder for testing data for dask cluster tests
this includes the intorduction of a testing framework for additional
tests with dask jobqueue clusters in the future.
Using the new test class will make it much easier to implement new tests
while using the correct test environment.
All tests should be implemented in this class or a derived class.

Unified the use of config_directories (remove the dask reimplementation)
this makes it easier for now an can be changed (simplify
config_directories) once the old workflows have been phased out.