feat: move all cluster settings in one dataclasse (dask and singularity)
all settings that are needed for a cluster have been moved into the
cluster configuration dataclasses.
Some options are only needed for singularity while others are needed for
dask only and some are needed for both.
this should be cleaned in the future.

Also many of the settings do not need be set in the input file again but
should instead be read from the config file directly.

Also the names of the tests are unified.