feat: start remote dask jobqueue cluster from .py file
We used to pickle the Python function to start the dask cluster on
remote.
For some reason this workflow did however not work when run from a
Macbook to a linux cluster.
Therefore this commit changes it to call a .py file remotely without
pickling it such that all arguments have to be send via the command line
call to start the .py file.
As a plus side, this new workflow exposes the code to start a dask
jobqueue cluster to the user such that he can also start a cluster
manually if wanted.