refactor: push expansion of parameters down to driver level
The creation of individual parameter dictionaries can be expensive,
for example, in the case of the expansion of random fields.
To decrease the compute they are pushed down to the Driver level, where
a parallel execution can be exploited.
Instead of expanding all field realisations in series in a for loop,
they are expaned within the Driver on Dask worker level.